+++ R/zzz.R
@@ -1,0 +1,6 @@
+# nocov start
+.onLoad <- function(...) {
+  vctrs::s3_register("waldo::compare_proxy", "plm")
+  invisible()
+}
+# nocov end