+++ tests/testthat/test-get-contributions.R
@@ -2,6 +2,8 @@ test_that("get_contributions() return authors' contributions", {
   df <- basic_df()
   aut <- Plume$new(df)
 
+  expect_s3_class(aut$get_contributions(), c("plm_agt", "plm"))
+
   df_contributions <- select(df, starts_with("contribution"))
 
   literal_names <- paste0(df$given_name, df$family_name)
@@ -11,22 +13,22 @@ test_that("get_contributions() return authors' contributions", {
   contributions <- condense(c(df_contributions))
   contributors <- lapply(list_initials, \(x) enumerate(dot(x)))
 
-  expect_equal_plm(
+  expect_equal(
     aut$get_contributions(),
     paste0(contributions, ": ", contributors)
   )
 
   contributors <- lapply(list_initials, enumerate)
 
-  expect_equal_plm(
+  expect_equal(
     aut$get_contributions(dotted_initials = FALSE),
     paste0(contributions, ": ", contributors)
   )
-  expect_equal_plm(
+  expect_equal(
     aut$get_contributions(dotted_initials = FALSE, divider = " - "),
     paste0(contributions, " - ", contributors)
   )
-  expect_equal_plm(
+  expect_equal(
     aut$get_contributions(
       role_first = FALSE,
       name_list = TRUE,
@@ -37,14 +39,14 @@ test_that("get_contributions() return authors' contributions", {
 
   contributors <- lapply(list_initials, \(x) enumerate(x, last = " & "))
 
-  expect_equal_plm(
+  expect_equal(
     aut$get_contributions(dotted_initials = FALSE, sep_last = " & "),
     paste0(contributions, ": ", contributors)
   )
 
   contributors <- lapply(list_initials, \(x) enumerate(sort(x)))
 
-  expect_equal_plm(
+  expect_equal(
     aut$get_contributions(alphabetical_order = TRUE, dotted_initials = FALSE),
     paste0(contributions, ": ", contributors)
   )
@@ -54,7 +56,7 @@ test_that("get_contributions() return authors' contributions", {
     enumerate(na.omit(x))
   })
 
-  expect_equal_plm(
+  expect_equal(
     aut$get_contributions(role_first = FALSE, literal_names = TRUE),
     paste0(literal_names, ": ", contributions)
   )