+++ NEWS.md
@@ -6,6 +6,8 @@
 
 ## Minor improvements and bug fixes
 
+* plume methods now print outputs in the same way.
+
 * New function `plm_push()` that can be used to insert author information into a Quarto document from a separate R script.
 
 * `get_author_list()` now accepts `format = ""` to return author names only. This is equivalent to `format = NULL` (#3).